refactor: move `paintHotspot` to separate module
